added inital project files
In this project, you will create an interactive photo gallery using
JavaScript and jQuery. Thumbnails and photos will be provided with
descriptions. At the top of the page you'll have a search area where
photos will hide and show depending on user input. When the user clicks
on a thumbnail, the photo will display in a lightbox. There should be a
back and previous arrows to cycle through photos.